fix : X축 스크롤 클릭 문제 해결
